requirements.txt
requests>=2.28.0
python-arango>=7.3.0